Update masks to reflect new working draft
The posts for css masks doesn't reflect the fact that the feature has gained new life with the advent of the W3C working draft. I suggest changing the status to "caution", but even if you guys are hesitant to do that, I hope you will at least alter the text to reflect the feature's new status (and hopeful future).