Replace `addEvent` function by `EventTarget.addEventListener`
As we'll be dropping support for IE8, there is no need for a `addEvent`
utility function anymore. In the past that function was used to garantuee
the event handler gets attached using `EventTarget.addEventListener` in
browsers that support it or using the `attachEvent` method in browsers
that lack support.

Removed the function declaration and replaced it with the
`EventTarget.addEventListener` method.