Update localstorage.md
When using localStorage, I had to learn that Safari Mobile limits the space available to zero bytes: Every write-operation fails. The polyfill by Remy doesn't work here either, since the Storage-object is actually available and thus the polyfill will not be used. So you need a fallback: Try to write to the localStorage. If this does not work, use a cookie instead. I would provide the sources, but my code's license does not allow me to do this.